added helper class. Divided tests into RoomServiceTest and ImportRoomsFromFile
